requirements.txt
ccxt
pandas
flask
python-dotenv
twilio
